add delete user

+++ app/controller/UserController.php
@@ -107,6 +107,7 @@ class UserController extends Controller
     {
         $rawData = file_get_contents('php://input');
         $data = json_decode($rawData, true);
+
         error_log(print_r($data, true));
 
         if (json_last_error() === JSON_ERROR_NONE) {
@@ -117,15 +118,16 @@ class UserController extends Controller
             $newName = isset($data['firstname']) ? $data['firstname'] : '';
             $newAdmin = isset($data['isAdmin']) ? $data['isAdmin'] : '';
 
-            list($username, $email, $surname, $name, $isAdmin) = $this->sanitizeUserInput($newUsername, $newEmail, $newSurname, $newName, $newAdmin);
+
+            list($username, $email, $surname, $name) = $this->sanitizeUserInput($newUsername, $newEmail, $newSurname, $newName);
 
             $userExists = $this->userRepository->readUserByUsername($username);
 
-            if ($userExists) {
+            if ($userExists && $oldUsername != $username) {
                 echo json_encode(['statusMessage' => 'duplicate', 'message' => 'Nutzername bereits vergeben']);
                
             } else {
-                $this->userRepository->updateUser($username, $email, $surname, $name, $isAdmin, $oldUsername);
+                $this->userRepository->updateUser($username, $email, $surname, $name, $newAdmin, $oldUsername);
                 echo json_encode(['statusMessage' => 'success', 'message' => 'Nutzer erfolgreich aktualisiert']);
             }
             
@@ -137,18 +139,5 @@ class UserController extends Controller
     {
         $users = $this->userRepository->readAllUsers();
         echo json_encode($users);
-    }
-    public function deleteUser()
-    {
-        $rawData = file_get_contents('php://input');
-        $data = json_decode($rawData, true);
-
-        if (json_last_error() === JSON_ERROR_NONE) {
-            $username = isset($data['username']) ? $data['username'] : '';
-            $this->userRepository->deleteUser($username);
-            echo json_encode(['statusMessage' => 'success', 'message' => 'Nutzer erfolgreich gelöscht']);
-        } else {
-            echo json_encode(['statusMessage' => 'error', 'message' => 'Invalid JSON data']);
-        }
     }
 }
